End transitions before starting new transition.
This ensures that the new transition starts from a good state so that
start-end state are computed corretly. Otherwise, it is possible that
the new animation starts from a transient state that doesn't compute
correct start and end states.

Some other cleanup:
 - TransitionListener that sets the regular digital clock to GONE is
   added before starting the transition to eliminate a race condition.
 - Setting the view to the beginning visibility is performed in
   AnimatorListener#onAnimationStart.

Bug: 132629519
Test: manual - flick to advance tracks while listening to YT music. It
is necessary to flick during the transition from lock to AOD to
reproduce the issue.

Change-Id: Id9d9732028a1693389f65dbb2f5c38ea60433bc6

Former-commit-id: b66e982fa7b4f8ce1c47c414195d1bdd93c8bcb0