Make package whitelist the last check in activity starter.
We should check all other conditions when deciding whether a background
activity start is permitted before checking the temporary package whitelist.
This will avoid confusion when examining bugreports where today an activity
start may appear to be permitted solely on the basis of the whitelist, when
in fact this is not the case.

Bug: 132871420
Test: manual
Change-Id: Ia5b6516076c72ae82d9a9c1c80613c4ef63b62cb

Former-commit-id: 44f0833d7820267e209c15601b3f3829f810258b