Always start launcher when empty home stack is being resumed
If home stack becomes empty for some reason and we're trying to find
and resume an activity in it - make sure to start home instead of
looking for the next top focusable stack. This will make sure that
home stack won't be left empty if a different display is currently
focused or higher in z-order.

There may be no activity in a home stack if launcher crashed, or at
boot when it switches between different activities.

Bug: 131440583
Test: atest WmTests:RootActivityContainerTests#testResumeFocusedStacksStartsHomeActivity_NoActivities
Test: atest WmTests:RootActivityContainerTests#testResumeFocusedStacksStartsHomeActivity_ActivityOnSecondaryScreen
Change-Id: I955d7d4827b2598f9c98cb30ec3fedac4e403a9d

Former-commit-id: 3ff5a0b44df949890a69d2451f8f057097655d09