Add flag for staged session to get ready during installation
Currently, the PackageManagerShellCommand returns success when a
staged session is successfully commited, regardless of the result
from pre-reboot verification. The flag will force the adb shell to
wait for the verification to complete and output result accordingly.
This will ensure a better experience for the testers.

Bug: 130039562
Test: manual testing by installing apex and observing new detailed
      success message
      $ adb install --apex --wait-for-staged-session-ready
      --wait 2000 com.android.tzdata.apex
Change-Id: Ib7aa9b5dfb683956e48e55461b29278069573aec
Merged-In: Ib7aa9b5dfb683956e48e55461b29278069573aec

Former-commit-id: f49c85f24b57cd167e92c36ffff5c9fc28759277