Fixed layout of quick_settings_header_info
The layout now allocates as much as it needs to QSCarrierGroup, and then
allocates leftover space to the status info Views.

The containing View (QSHeaderInfoLayout) splits the width between both
status (if both are visible) by allocating as much as they need. If the
total amount needed is more than what is available, it splits equally
unless a view requires less than half.

Test: Visual, using multiple Display sizes and Font sizes and Japanese
language (for some long strings)
Fixes: 132742416

Change-Id: I9f8ce68fa25fbe7d00451e59503575a07c3fd5f4

Former-commit-id: ceb057b6ed3f368d3dc52f5bbfebe199f9f7a925