Fix dropped rotation events when unlocking orientation
When unlocking orientation on the same frame as a rotation takes place
(example: orientation is locked to landscape, but phone is physically
oriented to portrait), onConfigurationChange is not invoked, and so
Global Actions is positioned as if it were in the locked orientation,
even though it has been rotated.

We work around this by post()ing the rotation unlock, so that it
happens on the next frame, at which point onConfigurationChange is
correctly invoked and the layout is re-oriented properly.

Fixes: 132581161
Test: manual
Change-Id: I1c11844e24bea115f9f44560fef8db863d19d7af

Former-commit-id: e05817ff03b9274a52e6d1b54d18576badfb0355