Move config for inflateSignalStrength
This moves the config files for determining the inflateSignalStrength
boolean into frameworks/base/core/res, and adds a utility method in
SettingsLib for retrieving it. This is needed so we can display
per-subscription signal strength icons in the DSDS header on the
Settings Network & internet page that match what SysUI shows.

Bug: 128855095
Test: manual, existing tests should pass
Change-Id: I57032f022439b2de3f7bf57757b2a49dc7eccfe1

Former-commit-id: c026de3dbf9c26e9f5ea942280191083cfec9314