Adjust down time to prevent subsequent modules being misleading
MagnificationGestureHandler cache some events to see if users want to
trigger magnification. If no magnification is triggered we inject the
events with original down time. This causes the subsequent modules,
like GestureDetector, be misleading. For example, injecting a cached
ACTION_DOWN event may mistrigger a long click if the down time of the
event is long enough ago.

Bug: 131815497
Test: a11y CTS & unit tests
Change-Id: I19508e8ef7c91af02cdc628463d5715f6de22cd3

Former-commit-id: 4a318ed2e84b47e3125f0db60a8fa36eab72597f