Refactoring: Rewrite links in Metalava instead of carrying patches.
Another CL in this topic introduces the metalava feature flag
--replace-documentation <ancestor java packages> <regex> <replacement>
to perform regex replacements in the documentation of some elements.

This CL uses that feature to re-write links in libcore/ojluni/
documentation that previously carried patches.

Since .bp files don't appear to support String literals that are not
(un)escaped, \ and $ occurring in the regexp/replacement parameters
need to be escaped in the .bp file.

Bug: 129765390
Test: Manually checked that this CL topic does not affect the result of
      "make ds-docs offline-sdk-docs"

Change-Id: Idb7366530c4ab255c1348cf48881cf2bde31be1e

Former-commit-id: 8f55941ebcce7cd0e306a3f660b4101c3b2fe9b0