Don't set as IME target when the stack is minimized in split screen
It could be reasonable that when in split screen mode, if navigating
to home screen by tapping back key or home key to let the docked primary
stack minimized, even the last state is focused on EditText, the IME window
should not be shown on the bottom or the divider handle icon hidden.

Added a check in WindowState#canBeImeTarget to ignore when the stack is
in docked minimized state.

Fix: 133151818
Test: atest WindowStateTests#testCanBeImeTarget
Test: manual
1. Open few apps from Apps list. (in this case Settings & Gmail)
2. Dock both the apps in split screen.
3. Tap on Settings search to bring up IME > Rotate the device >
   Switch IME focus to Gmail in lower frame.
4. Again rotate the device back to portrait > Bring IME focus to lower app
5. Navigate to home screen by tapping on back key and observe.
6. Expected the divider handle icon should be shown on the top of divider,
   and IME window should not shown on the bottom.

Change-Id: I1433f2766150d0916cbe08a0698c232fb7b4199c

Former-commit-id: f948afe6bcee51650c6b0057359669cbc5520bb0