Fix a bug about missing onConfigurationChanged
The original logic sends the new configuration to the non-activity
components only if there is any public field changed. However, it
doesn't use the latest configuration to compare to the current one,
which caused ActivityThread dropping some onConfigurationChanged
callbacks.

This CL uses the latest configuration to compare to the current one.

Fix: 132653657
Test: Steps in the bug
Change-Id: I969ef189c36ba1903503d8b7de5641103aed8cf3

Former-commit-id: 968064bed0a43b019cb7d84bed7493f6f78fda5e