Clear AppOpsController  when it stops listening
If AppOpsController stops listening, the lists were not cleared and it
may not receive the update to remove an active AppOp. To make sure that
we do not show outdated behavior once the controller is listening again,
we clear the lists on setListening(false).

When the controller goes back to listening, we will not have information
of anything that happened while the controller was not listening.

Test: atest AppOpsControllerTest
Fixes: 133263799
Change-Id: I2827b03d39d3b16c9cf6593537f7c1e30d7946f5

Former-commit-id: a4b4184f237ce7dab92d5c5db7367cc75e9e39e2