Clear VISIBLE_FOR_USER flag when isHiddenFromUser is true
The current code was clearing the LEGACY_POLICY_VISIBILITY flag when
isHiddenFromUser was returning true. This was causing issues since there was no
path to add that flag when the Window was visible for that user. The correct
way is to clear the VISIBLE_FOR_USER flag since that flag will get added again
when the window is visible for that user.

Fixes: 133235040
Test: Switch from one user to another and back. Launcher is visible.
Change-Id: I5161dd83a7a0811a7e3a6ab2af439f352f967a6c

Former-commit-id: bb6bd8ac180cdb8a87925fcf44ed8dfc99a85cbe