Add multi-display support for boot animation
This shows boot animations to the primary and other displays, verified
with 3-display setup.

A guest property, boot.animation.displays, controls which additional displays
to show the boot animation, in addition to the primary display. Use comma
as separator to specificy multiple values, for example:
  setprop boot.animation.displays 19260422155234049,19261083906282754

An example CL on how to set this system property: ag/9269083

Bug: 132892728

Test: adb logcat -s BootAnimation
Change-Id: I1ce3ebd57817246a7a0de42af9465e2a3517826b

Former-commit-id: f97c9e241fb3237e8f3cf13db8af5f12ca0910a8