Check system_server PID in NetworkStack calls
Add a check that callers with UID 1000 always have the same PID. This is
a proxy for checking that no system is designed to bind to the network
stack unless it is the system_server, as otherwise either the
system_server would start crashing, or that system would not have access
to binder calls.

Also remove access from PHONE_UID as it is not being used.

Test: Flashed, WiFi working, Bluetooth reverse tethering shows no
      permission issue.
Bug: 133209255
Change-Id: Ib848aaaedfd599c1d4437378846c7dda74352019

Former-commit-id: 4895c50e4f8ba9ec76c01e0e3d177c2a04a78fc0