Add an option to allow background starts for am broadcast.
Add an option to "adb shell am broadcast" to whitelist the broadcast
receiver to perform background activity starts. We could do this
unconditionally, but it's useful to be able to test the behavior with
and without the option.

This is gated on android.permission.START_ACTIVITIES_FROM_BACKGROUND,
which shell already has.

Bug: 132439856
Test: Manually send broadcast with and without the flag.
Change-Id: Iccd16e58c4f175bf3f5f1151dbe2bbf539489c33

Former-commit-id: c76767c21c85688335ac151eae666359981e8d3c