Not rendering the dark icons with the shelf anymore
Instead it is rendered by a dedicated container.
This will help with a lot of the transitions that
are necessary for the bypass and also fixes the
coloring of the notification icons on the lockscreen.

Fixes: 133869048
Bug: 130327302
Bug: 133182504
Test: atest SystemUITests
Change-Id: I422b3da488fdac6577173d41b26c67f81c7826c5
(cherry picked from commit 78d74b9765d2fe4919109f1c218f910e16009eaf [formerly 195dfc5db60fc913d060a692dc4ee7b0b02ece68])

Former-commit-id: 31292aad8afce9f01eb7c8900aaccde29df90e88