Make PiP window visible when user launches PiP in landscape mode
Two PiP animations are triggered when launching PiP in landscape mode
(one when first swiping up (FADE_IN), the other moves the PiP window to
portrait position (BOUNDS)). When screen rotation happens after the first
animation is started, it cancels it while leaving the window alpha to
close to 0. In order to make sure the window after rotation has the correct
alpha, we should set the alpha of the window when mCancelCurrentBoundsAnimation is true to 1.

Fixes: 133711307
Test: Play a YouTube video in landscape mode, and then swipe up. Make
sure that PiP window shows up after screen rotation forced by launcher
Test: atest PinnedStackTests
Change-Id: I3ec186b0e1e3f50dc91cfb03fc3847ddd7b5faef

Former-commit-id: 316b1f5028faa4a5775249d23f163269d15b3afc