Clean-up places that try to read wrap property.
Create a global function that gets the wrap property so everyone uses
the same definition.

It also fixes a bug in ProcessList.java that always set the using wrapper
variable to true since SystemProperty.get() always returned an empty string.
When this is fixed, the function getInputDispatchingTimeout() in
ActivityTaskManagerService.java will return 5 seconds instead of 60 seconds.
Therefore, there is a chance there could be timeouts occurring that didn't
occur before.

Bug: 133515802

Test: Built on a non-USAP device and verified that the wrap property works.
Test: Built on a USAP device and verified that the wrap property works.
Test: Ran the CtsWrapWrapDebugTestCases (verifies wrap.sh still works).
Change-Id: Ia3a85b6572aca6aafaf1405f8d3d97f4bb2441f6

Former-commit-id: 2d030bc4ddf3f54796b6526b3ac2a4d968a5cbb5