If window isn't requested to be visible, don't consider it
For the -1 overlay in Launcher, we'd like to keep it invisible
to reduce global system contention, to avoid allocating a surface
when just swiping up. However, this doesn't work yet as we always
consider the window as interesting even though it requested to
be not visible.

To fix that, we only consider a window as interesting if it has
also requested to be visible. Note: This doesn't apply for the main
window.

Test: go/wm-smoke
Test: Swipe up with special home APK
Bug: 123539527
Change-Id: I19f712051b51fed0435d80442ecb7bd9b0a69f80

Former-commit-id: 995e106ad775da3bb77bd9b25f4e014cdf79059c