Add test for rollback of userdata with multiple rollbacks
This test ensures that if a staged install is abandoned, and then
staged again, the userdata snapshot will match the correct rollback.
Since there will be multiple rollbacks in the ENABLING state for the
same package, it is important to ensure that the userdata for the
abandoned rollback is not used.

Test: atest StagedRollbackTest
Bug: 134035054
Change-Id: Icae7226dfb393c5ffc31dd0116f61141e84f5a42

Former-commit-id: 99ba7e9ca84fbcbf93d6176af029409b7ac7e1f4