Allow activity view to be max height
When we had the bubbles at the bottom we had to account for PIP dismiss
height but now they're at the top. Code can be simplified a lil now that
we're not bottom aligned.

Also adjusts the arrow to be invisible rather than gone which prevents a
small "jump" in the UI first time you expand a bubble

Fixes: 134155107
Test: manual - create a bubble, adjust slider to be 100%, observe that
               activity view height is as big as possible with space for
               "manage" button and bottom navigation (test gesture nav on
               and 2 or 3 button on)

Change-Id: Ifb9d3e23a6a0daa53372dca14ddba0cda8007952

Former-commit-id: a96c9ed4b0e6caa6883aa3df6fcaf453c284c8ae