Fix default home detection logic
Default home detection did not perform priority analysis if the default
home was null. Furthermore, the default home was retrieved before it was
initialized. By listening for boot complete and package changes, we
ensure that we get the correct default home.

This fixes incorrect handle behavior and metrics logging.

Test: Tested locally
BUG:135150547
Change-Id: I4ca36cbb1fad73905344e2434592a1893c178e23
FIX:135150547

Former-commit-id: 8c78deec28e39b24ec9655d6113ad387c6c455c7