Don't update dark boot setting onStart
Checking isPrimaryUser causes a delay in startup.  It's
unnecssary to check the theme on startup and on loading
the setting from settings since we only need the primary user's
dark theme setting.

Now, when persist.sys.theme is empty, we assume this means
to show the light-theme background (same as persist.sys.theme = 0/1).

Test: manual
Test: atest google/perf/boottime/boottime-test -v
Fixes: 135703697
Change-Id: I296eb2492dacf2273d204acc9cc84fc98c0e8827

Former-commit-id: d8f48bec045d20c95309e48f5e11cf31180f8d85