Fallback to default volume handling when session is released
Volume key events can be ignored when the foreground activity has
specified the MediaSession for handling media/volume key events
(via Activity#setMediaController()) but the session is released
(i.e. destroyed). Such enforcement was introduced not to use the
released session's information when picking the volume stream for
control.

This CL relaxes the behavior by falling back to the default volume
handling logic when the session is released.

Bug: 135966139
Test: Tested with the problematic app in the bug, and also with my own
      test app.
Change-Id: I976b0e8d18976f6cfdc3303cabdda55d2a101b4f

Former-commit-id: f3e269ce2155e5c254b99123f8e151edbe7b6aa2