Use the real activity component to find reusable task
The resolved target component was not save in task intent when an alias
activity was started. When starting alias activity again, the target
component was different from the alias component name which stored in
the task. So, we were unable to find a reusable task if the task was on
other display.

Bug: 135738129
Test: atest ActivityStackTests

Change-Id: I87249b6909eab9cd381f9de01b9c0edb32f2ae70

Former-commit-id: b05c8d8a1f42503205a71ec8193484a6ec5fdd91