Remove SurfaceControl#getHandle from JAVA apis 2/2
Holding on to a reference of the handle in Java will keep the server-side surface alive until
the reference is removed by GC. This may cause surfaces to be kept alive longer than necessary.
Instead hold on the surface control and call SurfaceControl#release which will release the local
reference to the server-side surface.

Bug: 136004147
Test: go/wm-smoke

Change-Id: Iff7d48ae1593cea2e188aa07c417f4c5a98887d5

Former-commit-id: bc9beab8fef8d266fbf08ed8fcfe1a468c863ead