Remove redundant appInfo field in ActivityRecord
ActivityRecord#appInfo contains duplicate information already present
in ActivityRecord#info. Since it can be updated at runtime, it needs
to be non-final and we need to remember to always update them
together.

This CL removes the appInfo field and switches all usages to
info.applicationInfo instead. No change in logic.

Test: Build and flash
Change-Id: Ia637a5874d2ec1f5516f7fb23e00c5a99e11f0a9

Former-commit-id: eceebbfa69cbf9b6d54aaddbba6d80804ad0079c