Add support to preload recents activity in background
The recents activity may be died in background for some reasons.
Then there may have jank when swiping-up the recents because it
needs to initialize at the same time. Especially when the recents
is the home activity. The case of configuration change is similar.

This CL provides the ability to start the recents activity in
background without affecting the foreground applications.

Bug: 127350205
Test: atest RecentsAnimationTest#testPreloadRecentsActivity
Change-Id: Idef72ce986eb910b302953705365437bc5ec2bf8

Former-commit-id: 609a8e29e7b1446b5e4039fc1d6a57aada5a82f3