Call syncInputWindows after calling transferTouchFocus
Since transferTouchFocus is invoked through SurfaceFlinger in a
transaction, it may not send the request to InputDispatcher until after
the input channel is removed. This was happening when the window was
getting removed right after the drag was started, causing the
transferTouchFocus to fail.

By calling syncInputWindows, it ensures that the transferTouchFocus is
sent to InputDispatcher before the channel could be destroyed.

Fixes: 133472746
Test: Steps in bug
Change-Id: I01d5e18a3fdddd8a1cc571e5df8101eec368838f

Former-commit-id: dc5ca62c0d6dffd88d2f8de0aa8f9d27d2de31a3