Fixed a few bugs with remoteInput
Because of our refactoring, we would crash whenever the
user clicked on reply while pulsing and not bypassing.
Also, the user would be sent to the shade when replying
from the bypass hun or smart replying, which is wrong.
Last, the logic wasn't working when the user didn't
have security.

Fixes: 67475697
Fixes: 135958017
Test: receive HUN on AOD and other surfaces, observe normal behavior
Change-Id: I68163ccfaa2dc2d998b2d68b29168e369670fb17

Former-commit-id: d17b3506577356f0e7d7a10b211613a97507da02