Fixed canceling fill request caused removal current session
A quick switch focus causes the fill request to be canceled and the
current session is deleted when it is considered to be a failure.
The new fill request will not work after that. To prevent this issue,
this change is to avoid cancelng request is treated as a failure.

Fixes: 135486325
Test: atest CtsAutoFillServiceTestCases
Change-Id: I213f79da5b248e46f5be1a940ceebb7ee99e00f3

Former-commit-id: 72c3f2bf276b4344e0fb3c29f70a9c3cfca5d75d