Remove Handler for Syncing Config Changes
Because WindowManagerService (WMS) and ActivityTaskManagerService (ATMS)
are under a single lock, there is no need to post to a handler within
WMS to prevent deadlock. This CL replaces the code that posts to the
handler with the code that gets executed when the handler receives the
message.

In addition, sendNewConfiguration has been wholly moved to
DisplayContent, so all callers of
WindowManagerService.sendNewConfiguration now call
DisplayContent.sendNewConfiguration.

Bug: 130807132
Test: go/wm-smoke
Change-Id: I278c4c5798983ce367014b9bbee07c4969bfce65

Former-commit-id: 6472e0ed138db93ed6c275dd1173804ae5125f3c