Read the high refresh rate blacklist from DeviceConfig.
When available, we should read the high refresh rate blacklist from
DeviceConfig, and use the device's configured list as a default /
fallback. This lets us change the blacklist in the field in case we find
additional compatibility issues, or one of the developers fixes their
app and asks us to remove them from the blacklist.

Fixes: 132338514
Test: atest WmTests:HighRefreshRateBlacklistTest
Exempt-From-Owner-Approval: TBR
Change-Id: I984dab60eeafff01758ac38c3db6056b33f662a6

Former-commit-id: f40ed275ef34e64119c98e1bc98e1cc0e6482621