Ensure permissions granted for correct SIM
This change ensures that the content URI permissions when writing or
reading an MMS are set for the carrier services package for the correct
SIM. Previously we did not specify the SIM, so the permissions would
only be granted to the carrier services package for the default SIM
(which would be whichever SIM is set to default voice).

This is basically a follow-up to b/128542685, which fixed the issue in
MmsService.

Fixes: 133514914
Change-Id: I3190961d53a72c3c14c453fcc32503ed579f5959
Test: Manually verified

Former-commit-id: 4cfd01d5dffec82d02992cdcb3646292accf3230