startop: Re-enable iorapd presubmit tests
This works-around the issue of the device rebooting by using shell
commands in DeviceSetup.

The downside is that this leaves the device in a semi-bad state
(there's no way to run shell commands *after* a test to restore
the regular iorapd). We assume it's ok because the TF harness
will reboot the devices for each new type of test.

Bug: 136200738
Bug: 134774708
Change-Id: I0e5efad4a1c76c5abc89c3bec77e2c6babcba261

Former-commit-id: be04ce1b77f34bf15306cc43b0fad85644d929d9