Refactor requestRefreshConfiguration call
Currently, it appears that requestRefreshConfiguration can only ever be
called with a single configuration value. That means, we can probably
switch to using an enum class for the different configuration values,
and simplify some of the existing logic.

But it's not super obvious from code that only 1 bit is set in this
"changes" variable. Refactor it a little to make it more obvious.

Bug: 137212522
Test: none
Change-Id: I8d21e7ea05e0271229f72354888b722dfdaf3929

Former-commit-id: 7507d7bc02908e4b2989559c8e9fd1c11fbfe7f0