Allow the vendor specific initialization of display_settings.xml.
In hawk_md, I tried to copy the intial settings in vendor-init: ag/7309059
Later, I found that this way doesn't work in SELINUX, since vendor-init is
not allowed to access core_data_files: http://shortn/_h6LOOHBVCF
So, I'd like to add a hook that vendor can provide the intial display
settings.
This is a stop-gap solution until b/123226552 is resolved.

Bug: 137233447
Test: Check if /vendor/etc/display_settings.xml is copied into /data/system/.
Change-Id: Ib87219e62c3414f26d63f072a0e767287ddb156e

Former-commit-id: f646144ab21a9f75e3e3a2aa2595d34d6dddcb4e