New APIs for SMSC address
Added getSmscAddress() and setSmscAddress() to access SMSC address in
(U)SIM through the SmsManager.
getSmscAddress() requires that the calling app is the default SMS app,
or has READ_PRIVILEGED_PHONE_STATE or the carrier privileges.
setSmscAddress() requires that the calling app is the default SMS app,
or has MODIFY_PHONE_STATE or the carrier privileges.

Test: Manual

Change-Id: Icb21aff450f71b30ef6a1181834014bf7b85b8e1
Signed-off-by: Taesu Lee <taesu82.lee@samsung.com>

Former-commit-id: f7488d087441fafd1279f3e6158cd3d197a9c5e3