Access ServiceConnector.mService in a thread-safe manner
binderDied() doesn't seem to have a defined thread on which it's called,
yet we want to abort queue processing asap on receiving it, so we have to
maintain state, modifiable from different threads.

The attached bug suggests modifying mService from a different thread is already
the case, so this just cleans up its usages to behave correctly in
parallel access scenario.

Fixed: 137744044
Test: presubmit
Change-Id: I9a09a69b5c1a1cd448a925bd7f8b2c6fa24357aa

Former-commit-id: 05bc95949b0329d9afce3da06ab0274ea23b2f61