Fixed thread leak in OomAdjuster during tests
OomAdjuster creates a thread in its ctor which is okay during normal
operation since we only call the ctor once. However, during tests it is
called per test setup, so we end up with a memory leak due to the
continuous creation of the thread.
We fix this by passing in the handler thread we use during testing into
the ctor since there is no need for a separate thread.

Test: presubmit
Bug: 137879065
Change-Id: Ia5664b3a4aa433a9fd55b714cc2a556c4e1b4eaa

Former-commit-id: ce8db89ce529343262f7f4c669dd2d8976afcff6