Fix issue with media notifs being misbucketed
Previous code assumed that "isHighPriority" == "is in top section",
which is not always true. Media notifs and some other notifs can appear
in the top section even if they're not high priority. Because we detect
section boundaries by iterating through the list until we find the first
notif where isHighPriority == false, we were sometimes drawing the
section boundary way too high. This change creates a new propery,
isInTopSection() that accurately tracks this state.

Setting this value in the proper location would require some seriously
destabilizing refactors, so instead we set it in the list comparator,
which is awful but here are.

Test: manual
Bug: 136709564
Change-Id: I19223720bac534ab92219a2962169097819e8efb

Former-commit-id: 8c1b763dcf5df6e7fc7177dad5e7390adb229fa4