Fixed that silent notifications weren't showing on AOD
Because we were listening to the tuning value but not
actually handling the null default value, the visibility
of the silent icons would be wrong by default.
Additionally, the whole settings check isn't actually
needed since we are already limited to the visible
notifications in the notification list and therefore
the setting is already automatically applied.

Fixes: 137977737
Test: atest SystemUItests
Change-Id: I93e23b08525c6bb8d2219a4bdaa0bbc64ce55d8d

Former-commit-id: 31ebb40789704c7a69c85a255bf401f47cf1a7f6