Delay idle alarm if there's an upcoming AlarmClock alarm.
The system won't go into deep idle when there's an AlarmClock set to go
off in the next hour. Knowing this, it doesn't make a lot of sense to
schedule an alarm to see if we can go into idle before the AlarmClock
alarm goes off. Changing to schedule the idle alarm to go off after the
AlarmClock alarm.

Bug: 135975445
Test: atest com.android.server.DeviceIdleControllerTest
Change-Id: I98bf4131b7b4a1006e37aa42bcbc7dda7e137d1d

Former-commit-id: ecf98fb4161f0ffa6a0060373c84cf4ac7e9f5c3