Support varargs for concatElements
concatElements currently can only support two arrays, this CL changes
it to use varargs so we can make use of it in an upcomming change which
needs a three way merge.

I've replaced the two way method with the varargs version, but I'm open
to discussion around keeping them both if the optimisations we can do
for a two way merge are seen as more valuable than the space saving from
having only one concatElements method.

Bug: 28437818
Test: atest ArrayUtilsTest
Change-Id: I634d279ceef7cd620efaf5a2c566e3ebadf6738e

Former-commit-id: 3c072e6fe6db9650f0a160647e8c90ab0db42fea