Removing JobScheduler heartbeat code.
Data shows that the quota based system has a small impact on battery life
while still addressing the core issues we had with the heartbeat system.
We can continue to tweak numbers and make some policy adjustments to the
quota system, but at this point, it appears safe to remove the
heartbeat code.

Bug: 138324538
Test: atest com.android.server.cts.JobSchedulerIncidentTest
Test: atest com.android.server.job.JobSchedulerServiceTest
Test: atest com.android.server.job.JobStoreTest
Test: atest com.android.server.job.controllers.ConnectivityControllerTest
Test: atest com.android.server.job.controllers.JobStatusTest
Test: atest com.android.server.job.controllers.QuotaControllerTest
Test: atest CtsJobSchedulerTestCases
Change-Id: I2b025ca426387ece123c0b8ed8bcc143f6c6d6db

Former-commit-id: 6c5dc642a15f333714caa7543b598d75cdabf41b