Fixed some issues when expanding from the launcher
We didn't exceed the touchslop automatically when
we got a touch from launcher which may have led us
to not expand when the user flung pretty fast.
Additionally could the velocity be very low if
we only had very few touch events, leading
to the panel also not expanding. We allow the panel
to expand easier in those cases.

This also fixed the vibration and the animation when
the fling happens really fast.

Fixes: 	137707746
Test: expand from launcher like crazy and observe that it does
Change-Id: Ie7d739f829a0c81352c86071f0e888ecc6c57108

Former-commit-id: 08bd435a91100c674d15d5c8759c576358df6685