Unset the notification expansion state when fully hidden
If the bouncer is showing after dragging down with locked
notifications and the screen locked, the visibleForExpansion
state wasn't properly reset. Even though we don't ask for
the bouncer when bypassing anymore, we're adding a
safeguard in case we change out minds. Whenever the notifications
are fully hidden, we unset the visibleForExpansion state.

Fixes: 	138261144
Test: drag down with bypass on, redaction and enforcing bouncer, turn off, then reach. No shelf should be visible
Change-Id: I2d237ce67da0e8fd3bf9e8d40e146f3a478ac5ce

Former-commit-id: 8a66e69242653f020d3e92701c89ad040a3754e4