Re-snooze Snoozed Notification
When Snoozing an already snoozed notification, the procedure is a NooP.
To fix this, we check if the notification we are trying to snooze is in
the snoozed notification list. We only do that if we cannot find the
notification in the displayed list (which was the only place where the
code looked for the notification). When we find the notification, we
update the snooze time.
Updated The code to look for snoozed group notifications
Bug: 70917141
Test: atest SnoozeHelperTest NotificationManagerServiceTest
Change-Id: I5ae17dcbc34ebccdcca6a64bb487fdef21703dd7

Former-commit-id: 24642da427ec9fda60d3a24428403a571e0b29b7