Make USE_BUGREPORT_API more widely available.
Previously this flag was in FeatureFlag for Settings, as only Settings
was using this flag. Now, we also need to use this feature flag while
migrating bugreports(from other callers such as Power buttons) to use the API.

In order to avoid flag duplication, adding it here, as Settings can
still access this feature flag.

Bug: 136548292
Test: Settings can still access this feature flag as before
Test: Builds and bugreports generated from Settings use the API flow
when the flag is turned on

Change-Id: Ia7409fc821e1438382e9b59c523ab72de519e861

Former-commit-id: 943c64b25ffc972cc75ac5b162a5b8c3197cf018