Clean up mockito references
Mockito has a leak that will OOM our tests otherwise.

Also mock out handler that was running on a background thread
because once the mock clean up code exists no mocks can
be used after the completion of the test.

Also in order to debug that I started making notification tags
more clear and I've left that debugging in for the future

Test: atest
Fixes: 138642368
Change-Id: Ib63d2035bd2476e3b57963b1aefb640a70d106cc

Former-commit-id: b6c837489b3a77cd28d246098be6640e9dc0f913