Fix application overlay visibility bug
System UI components may set the
SYSTEM_FLAG_HIDE_NON_SYSTEM_OVERLAY_WINDOWS flag in
WindowManager.LayoutParams to force application overlays to be
invisible. StatusBarViewController is responsible for setting this when
the user locks their phone, and then removing it when the user unlocks
their phone.

Currently there is a bug where, while StatusBar removes this flag, the
update is not propagated through WindowManagerService. This causes
application overlays to be invisible when they should not be. This
affects all application overlays.

Repro steps:

1) Enable rotation if disabled
2) Enable an application overlay (e.g. Live Caption)
3) In portrait orientation, open e.g. YouTube.
4) Rotate to landscape orientation
5) Turn off the phone screen
6) Rotate to portrait orientation
7) Turn on the phone screen and unlock the phone

Application overlays are now invisible. Rotating again fixes it.

This CL fixes the logic error that was causing this bug.

Bug: 133747592
Test: go/wm-smoke
Change-Id: Idab0629765cbe048fc1eb4413e529464cad876ff

Former-commit-id: 44742c0ef8608495599c874b26b34b6745d21000