PointerLocation should not show if display is in mirror mode.
If a virtual display is created without flag FLAG_OWN_CONTENT_ONLY,
and no window initialized, it will automatically enter the mirror mode.

So we need to exclude the pointer location window when checking the
display has content, to make sure it won't leave the mirror mode if there
is no other window exist.

Test: manual
Bug: 138487611
Change-Id: I60d70ea578df4b18b1666d439a2c925f3310351f

Former-commit-id: fbc8f413d1806d5104ac1424d586092d38b8a7ee