Only show the thermal shutdown warning when thermal reboot
showThermalShutdownWarning() is on the path where systemui starts
and restarts. Previously, we just checked the boot reason in
determining whether to show the warning, which makes it unexpectedly
seen when systemui restarts from crashes.

This CL caches the last boot count and compares with the current one
to ensure that there is a thermal reboot before the warning is shown.

Test: atest PowerUITest
Bug: 138089094
Change-Id: I0a51d00703d291ba10a6ab4f348308a14000a20e

Former-commit-id: 9b78673350e8edb1ce66edd7d764c05f0ae257c7