Fix inconsistent OomAdjuster computation result
In case of cycle dependencies, while one of the process within
the cycle gets bound by mutliple other processes, the procState
result would be inconsistent across runs. Fix is to pick the
minimum one of the raw and current scores prior to scanning the
service/provider lists in case of cycles-reevaluation.

Bug: 138673927
Test: atest MockingOomAdjusterTests
      cts-tradefed run singleCommand cts-dev -m CtsAppTestCases \
      -t android.app.cts.ActivityManagerProcessStateTest

Change-Id: I8f33c12c1625da6f336d79e44e91158c84beb6ac

Former-commit-id: caf294a2cef3fd821f2bbfd56414279465f81deb