Prevent SurfaceView from drawing over parent surface insets
Parent SurfaceView to a bounds layer that enforces a crop preventing it
from drawing over the surface insets. The bounds layer crop is set to the
surface insets and updated when the parent surface size changes or the
parent surface insets change.

The SurfaceView has a z order relative to the ViewRootImpl surface so
that it can be behind or in front of the main content. If the ViewRootImpl
surface changes, then SurfaceView takes advantage of the
ViewRootImpl.SurfaceChangedCallback to update the relative z order.

Bug: 132205507
Test: go/wm-smoke
Test: test preserve surfaces code path with split screen and ensure relative z is preserved
Test: test SurfaceView apps, YouTube, maps & camera.
Test: try to repro maps pip issue described in bug
Change-Id: I0b2a3612b7e12ba66abb39b6bad4968442743cdd

Former-commit-id: 8cb00aee22a4f82051c586bca82ccf41c71f60d6