Only collapse the stack onTaskMovedToFront if we're not animating
I think it should be safe to skip this if the stack is animating.

If the stack is animating due to collapse then it's fine because
we're collapsing anyways.

The situation to be concerned about is:
1) Tap stack to expand
2) Almost simultaneously tap app / something that would bring
   activity up

Step 2 can never happen because as soon as the stack is expanding,
bubbles is grabbing any touches -- so that second tap would be
piped to bubbles.

Test: manual - 1) have bubbles
               2) open any app
               3) swipe up to home and quickly tap bubble stack
               => bubbles are expanded as expected
Bug: 138941969
Change-Id: Ie5f12327f1ecf8801a9b800b81f505310e7f63fe

Former-commit-id: 047e24ec8ab37830ba0577b303e892ff7c3fdb05