Cleans up sharedUser regardless of creation logic
Prior to this change, we were not updating the shared user a package
setting belongs to if the package setting was not deep copied. In the
case that shared users didn't match, we were not deep copying and so
were not updating later. There is a more fundamental change to be made
here, but we're limiting risk by keeping this change as small as
possible.

Bug: 138919405
Test: atest AppSecurityTests
Test: manual; flash P then flash this build: observe consistent PackageSettings
Change-Id: Ib22f9e6a1fb3104bec4eb3e5b1de983e19d49dce
(cherry picked from commit 1abb09e2d6baf8f43902241741a4f128a66b182f [formerly cf9a4835c110b3b7c064e1dde97ffb38ef0424eb])

Former-commit-id: db5100ec6f11592c24fcba7dffcb26b6756e886c