Prevent QSPanel from remeasuring
Once QSPanel has been measured, do not measure it again in the pass from
QSContainerImpl.

This prevents measuring with a slightly bigger size
(why?) that causes it to recalculate the number of rows (and the number
of pages) of PagedTileLayout. In some cases, this was causing a
different measure while scrolling for an instant, changing the number of
pages and therefore sending the user to page 0 again.

Fixes: 135136427
Test: manual, following configs:
	* landscape and portrait
	* 1, 2 and 3 pages
	* Security footer off and on with 1 and 2 lines

Change-Id: I74e3ce68bd2d3a8b7af1c84d9d0164ce6a672ed1

Former-commit-id: 0b7d617af2616b221e4a5e1f796b481b2ff1cdce