Add a root layer in DisplayContent
To enable surface animation on the overlay layer and the windowing
layer, a leash needs to be created on top of them, but because they have
no parent, there is nothing to reattache them to after the leash is
destroyed.

Adding a root layer as a container layer enbale us to workaround this
issue.

Test: presubmit and manually with the new rotation animation running on
with SurfaceAnimator (in upcoming CL)
Bug: 131727512

Change-Id: I156a7a479a055341a67eafcc203ee2c83d88a30e

Former-commit-id: 1812c76943e71c2c7c9d319aac605232c86ec853