Pin staledataclass-annotation-processor to Java language level 8.
This target tries to use --add-modules and --add-exports to get access
to internal classes from the jdk.compiler module. This works for Java
code (unsupportedappusage-annotation-processor) but not for Kotlin
code, as here. The kotlinc invocation fails.

As a workaround for this issue, this change pins this to Java language
level 8, so the JPMS doesn't apply.

Test: EXPERIMENTAL_JAVA_LANGUAGE_LEVEL_9=true make staledataclass-annotation-processor
Bug: 139342589
Change-Id: I5234bd3daf5f5347547ead50445ba989b38cbf6e

Former-commit-id: 50c58e48adabf15a038a499459bae94a7e7f1744