Set a minimum height for invocation lights
The height of the invocation lights view was set to the corner
radius, so phones with square corners (and therefore corner radius
zero) had zero-height views (making the motion indication invisible).

This change sets the minimum possible height to the height of the
lights themselves (the stroke width).

Fix: 138667570
Test: invoked assistant before and after change on walleye, change
causes assistant motion lights/indication to be visible

Change-Id: I54e765601688abb8d5c736a9b68556a876912639

Former-commit-id: c874531e45fc20c2553499c5a59a86bf86b27bd0