Do not stop idmap2d after OMS initialization
The OMS was calling stopIdmapService after OMS initialization finished.
This method was not setting the cached service binder to null and the
service binder was being reused when an overlay is enabled or disabled
within the 10 seconds the connection to the service is still opened.

This change no longer calls stopIdmapService and makes startIdmapService
and stopIdmapService private to IdmapDaemon.

Bug: 139307780
Test: enabled overlay within 10 seconds of OMS startup ending and
      observed state STATE_ENABLED.
Change-Id: I46b6a3631278c4dbdf5e5d373e085bac5186ed9a

Former-commit-id: b8fbab621840616339a0831ddca07ce97a130cfc