Return correct value for Activity#releaseInstance()
The documentation of Activity#releaseInstance() API tells that it
returns 'true' if it "has started the process of destroying its
current instance". However, the actual implementation that calls
ActivityRecord#destroyImmediately() would only return 'true' if
activity was removed from history immediately. In most cases it
would return 'false', since it would first send "destroy" message to
the client and wait for it to report back.

This CL switches the system-server implementation to return 'true' if
the activity is either destroying or destroyed as a result of
calling Activity#releaseInstance().

Bug: 137329632
Test: atest CtsWindowManagerDeviceTestCases:ActivityTests
Change-Id: I7d7287261d1742cffcf11ffeb32475019b978348

Former-commit-id: f49a58c095e3c7e0cd6f1a42befcd34e88b8a7b1