Extract common display density configuraiton code
We're currently importing the whole of SettingsLib into the Settings
Backup Agent just to access a single static method. This seems to
have contributed to a the regression which relates to b/139373401.

This CL separates the 3 related methods out into their own build
target so the amount of code included in SettingsBackupAgent is
much smaller.

A later CL will remove the original location after all usages
have been redirected to this new location

Bug: 139373401
Test: atest SettingsProviderTest SettingsLibRoboTests
Change-Id: I36fbc21daddcf068491b58db01b66f540ac26cf7

Former-commit-id: e5e7916b06fadb1ea06ab6bb2df9f264d4e8022e