Fix bug in Time.getJulianDay()
Time is generally deprecated because it has various problems
including Y2038 issues. The static Time.getJulianDay() method
doesn't have Y2038 issues but it did have an issue with
negative millis values which is fixed in this commit.

See the associated cts change for improved tests.

Documentation has also been improved in this commit.

Bug: 16550209
Test: atest android.text.format.cts.TimeTest
Change-Id: Ia4092efc92a24cd5741b2a8dd75a0680008cce27

Former-commit-id: 5898cc1060941a8bec0021d8d4f7b35f71792cb5