Adds app enumeration support for all components
This change adds consideration for more than just activities when
computing matches between packages for app enumeration / filtering.

It also changes the failure logic when parsing a package to allow for a
queries intent tag that contains no action, but a scheme or one that
contains no data tag but one action (or both). Previously it would have
been impossible to enumerate an app purely based on the authority of one
of its providers.

Bug: 136675067
Test: adb shell device_config put package_manager_service package_query_filtering_enabled true && atest AppEnumerationTests
Test: atest AppsFilterTest
Change-Id: I07bb449e78fb79a2ed61f75b37e582e0f3467a2d

Former-commit-id: 0c9257b73c740b12f7dbcf7cb34d091b16452a3a