Ensure lock icon animates closed on long press
There's a bug in the current implementation of the lock icon that
prevents it from animating from the unlocked to the locked state
sometimes when it's long-pressed. This is purely an animation bug, as
the device remains locked and requires authentication. However, the
unlocked icon sticks on the screen in this state, leading to a
confusing UX.

Test: manual:
0. Ensure face is enrolled and "Skip lock screen" is off
1. Wake device and unlock with face
2. Long-press lock icon to re-lock device
3. Turn screen off and back on, unlock with face again
4. Long-press lock icon a second time

Before: "Device will stay locked" message shown, but icon stays unlocked
After: "Device will stay locked" message shown, icon animates to locked

Bug: 139831703
Change-Id: Ifc1fe46ccdac5e069004ddc49ffa9d17fd906506

Former-commit-id: c48eff5016cc45a43e3b6035aef5e0f258a08e3a