Add wrapper functions to eventually remove enum logic from callers.
The bugreport API uses newly added enum for bugreport options
(example: BugreportParams.BUGREPORT_MODE_FULL). So, now we need to
maintain and keep in sync both of these enums as requestBugreport will
use the bugreport API; Need to move towards one enum for bugreport
options.

Currently, the caller of requestBugReport(int bugreport) needs to know
the underlying enum for the bugreport it is requesting. Instead of
changing the enum logic at the caller of requestBugreport(), create
wrapper functions for specific bugreports.

Bug: 137825297
Test: Builds
Change-Id: I62a58511867624f9662287ee39f78a31ff9cd765

Former-commit-id: e9998513db50487b4e92b9d5b7d420e48b619ca0