Update visibility of finishing activities
A visibility of a trampoline activity may not be updated when the
activity it launches becomes visible in some cases. This CL makes
sure that the visibility of a finishing activity is updated to
'false'.

In a normal flow when a trampoline activity launches the next one,
ActivityStack#resumeTopActivityInnerLocked() sets the visibility to
'false' when the next one is being resumed. However, if the
RootActivityContainer#ensureVisibilityAndConfig() is called before
that, it can update the visibility of the top activity without
updating the visibility of the trampoline (visibility updates were
skipped for finishing activities).

Bug: 139808754
Test: Unit tests for ActivityRecord and CTS lifecycle tests
Change-Id: Idc89743fb708487594e5b56f2e860e4ffc78fd14

Former-commit-id: 149d9ad3fe1842e56d97a1f5050450b04a6ee3f5