Clear notifications deletes configs
When resetting the application data, the notification settings are reset
to disalowed/none even if the app should be allowed by default. As a
result, all packages and components that are supposed to be allowed by
default are disabled. The
meaning of reset app data should return to default instead of
disallowing/deleting the app data

Bug: 135767808
Test: Test: atest NotificationManagerServiceTest
Change-Id: Ida15d3887106c3ccf3f2fd3127d1b714211f2a99

Former-commit-id: 4204f2564de95b2b8b5aaca26a7fac972eba3a21